LayoutTests/http/tests/navigation/post-goback-same-url.html

<html>
<head>
<script>
    
if (window.testRunner) {
    testRunner.dumpBackForwardList();
    testRunner.dumpAsText();
    testRunner.queueLoad("resources/post-goback-same-url.php");
    testRunner.queueLoadingScript("submitForm()");
    testRunner.queueBackNavigation(1);
}
    
</script>
</head>
<body>
https://bugs.webkit.org/show_bug.cgi?id=25223 - Back button after form
submission to the same URL fails to navigate
</body>
</html>
