LayoutTests/http/tests/navigation/post-goback-same-url-expected.txt



============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/post-goback-same-url.html  **nav target**
curr->  http://127.0.0.1:8000/navigation/resources/post-goback-same-url.php  **nav target**
        http://127.0.0.1:8000/navigation/resources/post-goback-same-url.php  **nav target**
===============================================
